Extract scopes
