+++ app/models/csv_dumps/dump_scope.rb
@@ -1,0 +1,15 @@
+module CsvDumps
+  class DumpScope
+    attr_reader :resource
+
+    def initialize(resource)
+      @resource = resource
+    end
+
+    private
+
+    def read_from_database(&block)
+      DatabaseReplica.read("dump_data_from_read_slave", &block)
+    end
+  end
+end
