+++ app/workers/classifications_dump_worker.rb
@@ -12,47 +12,11 @@ class ClassificationsDumpWorker
     @formatter ||= Formatter::Csv::Classification.new(cache)
   end
 
-  def each
-    read_from_database do
-      completed_resource_classifications.find_in_batches do |batch|
-        subject_ids = setup_subjects_cache(batch)
-        setup_retirement_cache(batch, subject_ids)
-        batch.each do |classification|
-          yield classification
-        end
-      end
-    end
+  def get_scope(resource)
+    @scope ||= CsvDumps::ClassificationScope.new(resource, cache)
   end
-
-  private
 
   def cache
     @cache ||= ClassificationDumpCache.new
-  end
-
-  def setup_subjects_cache(classifications)
-    classification_ids = classifications.map(&:id).join(",")
-    sql = "SELECT classification_id, subject_id FROM classification_subjects where classification_id IN (#{classification_ids})"
-    c_s_ids = ActiveRecord::Base.connection.select_rows(sql)
-    cache.reset_classification_subjects(c_s_ids)
-    subject_ids = c_s_ids.map { |_, subject_id| subject_id }
-    cache.reset_subjects(Subject.where(id: subject_ids).load)
-    subject_ids
-  end
-
-  def completed_resource_classifications
-    resource.classifications
-    .complete
-    .joins(:workflow)
-    .includes(:user, workflow: [:workflow_contents])
-  end
-
-  def setup_retirement_cache(classifications, subject_ids)
-    workflow_ids = classifications.map(&:workflow_id).uniq
-    retired_counts = SubjectWorkflowStatus.retired.where(
-      subject_id: subject_ids,
-      workflow_id: workflow_ids
-    ).load
-    cache.reset_subject_workflow_statuses(retired_counts)
   end
 end
