+++ app/workers/subjects_dump_worker.rb
@@ -12,18 +12,8 @@ class SubjectsDumpWorker
     @formatter ||= Formatter::Csv::Subject.new(resource)
   end
 
-  def each
-    read_from_database do
-      project_subjects.find_each do |subject|
-        yield subject
-      end
-    end
-  end
 
-  def project_subjects
-    Subject
-      .joins(:subject_sets)
-      .eager_load(:subject_sets, :locations)
-      .merge(resource.subject_sets)
+  def get_scope(resource)
+    CsvDumps::SubjectScope.new(resource)
   end
 end
